clean: Stop using better-files and use java.nio directly
I want to start using Scala Native instead of Graalvm native-image.
`better-files` is not being maintained a lot these days.
Since we do little use of it in this library I ported the existing
code to use java.io and java.nio directly, so we have less dependencies
and the tools are easier to port to Scala Native